DockerImage/requirements.txt
python-decouple==3.6
requests==2.28.1
vk-api==11.9.9
